Remove some configoptional changes
