+++ confmap/internal/unmarshaloption.go
@@ -3,8 +3,6 @@
 
 package internal // import "go.opentelemetry.io/collector/confmap/internal"
 
-import "github.com/go-viper/mapstructure/v2"
-
 type UnmarshalOption interface {
 	apply(*UnmarshalOptions)
 }
@@ -12,19 +10,11 @@ type UnmarshalOption interface {
 // UnmarshalOptions is used by (*Conf).Unmarshal to toggle unmarshaling settings.
 // It is in the `internal` package so experimental options can be added in xconfmap.
 type UnmarshalOptions struct {
-	IgnoreUnused              bool
-	AdditionalDecodeHookFuncs []mapstructure.DecodeHookFunc
+	IgnoreUnused bool
 }
 
 type UnmarshalOptionFunc func(*UnmarshalOptions)
 
 func (fn UnmarshalOptionFunc) apply(set *UnmarshalOptions) {
 	fn(set)
-}
-
-// ApplyUnmarshalOption simply calls (UnmarshalOption).apply. This function allows us
-// to keep the `apply“ function private and therefore keep `confmap.UnmarshalOption`
-// without any exported methods.
-func ApplyUnmarshalOption(uo UnmarshalOption, set *UnmarshalOptions) {
-	uo.apply(set)
 }
